Add caching and change login test spec
